% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create the reproducible HTML sensory report}
\usage{
create_report(
  path_to_data_folder,
  output_dir,
  report_template = system.file("report", package = "sensometrics")
)
}
\arguments{
\item{output_dir}{Character. Path of the folder to save the report.}

\item{report_template}{Character. Path of the folder containing the report template.}

\item{date}{path_to_data_folder. Path to data folder.}

\item{session}{Character. ID of the sensory session.}

\item{products}{Character. Vector of products to include into the report.}
}
\value{
A list with the folder and the file of the created report.
}
\description{
Create the reproducible HTML sensory report
}
\examples{
# Create a temporary directory
my_temp_dir <- tempfile()
dir.create(my_temp_dir)

# Create the sensory report
report <- create_report(
  path_to_data_folder = "data/20240425",
  output_dir = my_temp_dir
)

# Open the sensory report
browseURL(
  file.path(
    report$output_dir_report, 
    report$output_file_report
  )
)

# Delete temporary directory
unlink(my_temp_dir, recursive = TRUE)
}
