% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_from_file_name.R
\name{get_date_from_file_name}
\alias{get_date_from_file_name}
\title{Get the date of the session from file name}
\usage{
get_date_from_file_name(file)
}
\arguments{
\item{file}{Character. Path of a file containing the date inside its name.}
}
\value{
Date
}
\description{
Get the date of the session from file name
}
\examples{
get_date_from_file_name(
  file = "data/data_info_products_20240425.xlsx"
)
}
