% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_from_folder_name.R
\name{get_date_from_folder_name}
\alias{get_date_from_folder_name}
\title{Get the date of the session from folder name}
\usage{
get_date_from_folder_name(path)
}
\arguments{
\item{path}{Character. Path of a folder containing the date inside its name.}
}
\value{
Date
}
\description{
Get the date of the session from folder name
}
\examples{
get_date_from_folder_name(
  path = "data/20240425"
)
}
